Time: 0 ms (100.00%), Space: 6.3 MB (71.72%) - LeetHub

+++ 0070-climbing-stairs/0070-climbing-stairs.cpp
@@ -7,6 +7,6 @@ public:
             b += a;
             a = temp;
         }
-        return a;
+        return (int)a;
     }
 };
